Move financial details to new controller
