+++ Hippo.Web/ClientApp/src/App.tsx
@@ -119,9 +119,12 @@ const App = () => {
               }
             />
             <Route
-              path="/:cluster/admin/financialdetails"
+              path="/:cluster/financial/financialdetails"
               element={
-                <ShowFor roles={["System"]} alternative={<NotAuthorized />}>
+                <ShowFor
+                  roles={["System", "FinancialAdmin"]}
+                  alternative={<NotAuthorized />}
+                >
                   <FinancialDetail />
                 </ShowFor>
               }
