+++ Hippo.Web/ClientApp/src/AppNav.tsx
@@ -105,11 +105,11 @@ export const AppNav = () => {
                   Cluster Admins
                 </NavLink>
               </ShowFor>
-              <ShowFor roles={["System"]}>
+              <ShowFor roles={["System", "FinancialAdmin"]}>
                 <NavLink
                   id="financialDetails"
                   className="nav-item nav-link"
-                  to={`/${cluster.name}/admin/financialdetails`}
+                  to={`/${cluster.name}/financial/financialdetails`}
                   style={({ isActive }) =>
                     isActive ? { fontWeight: "bold" } : {}
                   }
